refactor(app): add the Route group restructure project layout and move files to new directories
Move components and pages to (main) and (second) route groups for better organization
Update imports and paths to reflect new file locations
Delete old unused files and consolidate duplicate styles
