+++ src/app/(main)/news-list/[category]/layout.tsx
