+++ src/app/(main)/news-list/[category]/page.tsx
