+++ src/app/(main)/news/[slug]/ItemIdata.module.css
