+++ src/app/(main)/news/[slug]/ItemIdata.tsx
