+++ src/app/(second)/layout.tsx
@@ -1,0 +1,35 @@
+import type { Metadata } from "next";
+import { Roboto, Mulish } from "next/font/google";
+
+const Robotofont = Roboto({
+    variable: "--font-roboto",
+    subsets: ["greek"],
+    fallback: ["Arial"]
+});
+const Mulishfont = Mulish({
+    variable: "--font-Mulishfont",
+    subsets: ["latin"],
+    fallback: ["Arial"],
+    display: "swap"
+});
+
+export const metadata: Metadata = {
+    title: "News App",
+    description: "Generated by create next app",
+};
+
+interface Iprops {
+    children: React.ReactNode;
+
+}
+export default function RootLayout({ children }: Iprops) {
+
+    return (
+        <html
+            lang="en"
+            className={`${Robotofont.variable} ${Mulishfont.variable}`}
+        >
+            <body>{children}</body>
+        </html>
+    );
+}
